requirements.txt
torch==2.1
transformers==4.34.0
ctransformers
gradio==3.41.2
fastapi
uvicorn
python-dotenv
langchain==0.3.16
chromadb==0.5.0
pypdf
pypdf2
sentence_transformers
tiktoken
tqdm
together
langchain-together
langchain_community
pydantic==2.7.4